// monitorai-services-master/directus/schema.ts
export interface components {
    schemas: {
// ...
        Users: {
            /**
             * @description Unique identifier for the user.
             * @example 63716273-0f29-4648-8a2a-2af2948f6f78
             */
            id?: string;
            /**
             * @description First name of the user.
             * @example Admin
             */
            first_name?: string;
            /**
             * @description Last name of the user.
             * @example User
             */
            last_name?: string;
            /**
             * Format: email
             * @description Unique email address for the user.
             * @example admin@example.com
             */
            email?: string;
            /** @description Password of the user. */
            password?: string;
            /**
             * @description The user's location.
             * @example null
             */
            location?: string | null;
            /**
             * @description The user's title.
             * @example null
             */
            title?: string | null;
            /**
             * @description The user's description.
             * @example null
             */
            description?: string | null;
            /**
             * @description The user's tags.
             * @example null
             */
            tags?: string[] | null;
            /**
             * @description The user's avatar.
             * @example null
             */
            avatar?: (string | components["schemas"]["Files"]) | null;
            /**
             * @description The user's language used in Directus.
             * @example en-US
             */
            language?: string;
            /**
             * @description The 2FA secret string that's used to generate one time passwords.
             * @example null
             */
            tfa_secret?: string | null;
            /**
             * @description Status of the user.
             * @example active
             * @enum {string}
             */
            status?: "active" | "invited" | "draft" | "suspended" | "deleted";
            /**
             * @description Unique identifier of the role of this user.
             * @example 2f24211d-d928-469a-aea3-3c8f53d4e426
             */
            role?: string | components["schemas"]["Roles"];
            /** @description Static token for the user. */
            token?: string | null;
            /**
             * Format: date-time
             * @description When this user used the API last.
             * @example 2020-05-31T14:32:37Z
             */
            last_access?: string | null;
            /**
             * @description Last page that the user was on.
             * @example /my-project/settings/collections/a
             */
            last_page?: string | null;
            provider?: string;
            external_identifier?: string | null;
            auth_data?: unknown;
            email_notifications?: boolean | null;
            appearance?: string | null;
            theme_dark?: string | null;
            theme_light?: string | null;
            theme_light_overrides?: unknown;
            theme_dark_overrides?: unknown;
            telegram_id?: string | null;
            company?: (string | components["schemas"]["ItemsCompanies"]) | null;
            phone?: string | null;
            repair_team?: (string | components["schemas"]["ItemsRepairTeams"]) | null;
            holding?: (string | components["schemas"]["ItemsHoldings"]) | null;
            policies?: unknown;
        };
// ...
    };
// ...
}